Add support to establish more than one connection with a client
This change enables Hermes to establish more than one connection with the given server
and perform a round robin like way of distributing the requests between every connection.

This is interesting when, for example, requests are directed to a Network Load
Balancer which performs the load balancing for every request it receives. But when
the connections are persistent, like in http2, every request will go directly
to the same backend instance if we establish just one connection.